requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
Markdown==3.5.2
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
